+++ scripts/provision/debian/custom/basic/B_packages.sh
@@ -6,6 +6,6 @@ set -x
 
 apt-get update
 DEBIAN_FRONTEND=noninteractive apt-get -y upgrade
-DEBIAN_FRONTEND=noninteractive apt-get -y install tree vim bash-completion
+DEBIAN_FRONTEND=noninteractive apt-get -y install tree vim bash-completion apt-transport-https
 
 echo "Finishing ${0}.."