+++ CHANGELOG.rst
@@ -1,4 +1,5 @@
 [Unreleased] ([unreleased])
 ---------------------------
+* add script rex_install_git
 * improve provision scripts
 * initial version